+++ include/methcla/plugins/node-control.h
@@ -23,5 +23,6 @@ METHCLA_EXPORT const Methcla_Library* methcla_plugins_node_control(const Methcla
 
 #define METHCLA_PLUGINS_DONE_AFTER_URI METHCLA_PLUGINS_URI "/done-after"
 #define METHCLA_PLUGINS_ASR_ENVELOPE_URI METHCLA_PLUGINS_URI "/asr-envelope"
+#define METHCLA_PLUGINS_EXPONENTIAL_FADE_URI METHCLA_PLUGINS_URI "/exponential-fade"
 
 #endif // METHCLA_PLUGINS_NODE_CONTROL_H_INCLUDED