+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 * Fix bus zeroing logic (#102)
 * Add operator bool to `Methcla::NodeId`
 * Add `Methcla::Engine::nodeEndedHandler` API method returning `/node/ended` notification handler
-
+* Add `ExponentialFade` plugin (`METHCLA_PLUGINS_EXPONENTIAL_FADE_URI`) to `methcla_plugins_node_control` library
 ### 0.2.0
 
 * Rename `Methcla::Engine::freeNode` to `Methcla::Engine::free`